Structure refactor     * Moved app ui example to user_interface layer     * Moved all test to src folder     * Removed from travis package task due a long time process unnecessary for github     * Pending to see why is not working package task due a missing main.js
