--- app/store/configureStore.js
@@ -1,6 +1,0 @@
-// @flow
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./configureStore.production'); // eslint-disable-line global-require
-} else {
-  module.exports = require('./configureStore.development'); // eslint-disable-line global-require
-}
