+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.base.js
@@ -4,7 +4,7 @@
 
 import path from 'path';
 import validate from 'webpack-validator';
-import { dependencies as externals } from '../../../../../app/package.json';
+import { dependencies as externals } from '../../../user_interface/package.json';
 
 export default validate({
   module: {
@@ -19,7 +19,7 @@ export default validate({
   },
 
   output: {
-    path: path.join(__dirname, 'app'),
+    path: path.join(__dirname, 'user_interface'),
     filename: 'bundle.js',
 
     // https://github.com/webpack/webpack/issues/1114
