+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.development.js
@@ -19,7 +19,7 @@ export default validate(merge(baseConfig, {
   entry: [
     `webpack-hot-middleware/client?path=http://localhost:${port}/__webpack_hmr`,
     'babel-polyfill',
-    './app/index',
+    './src/openMarket/user_interface',
     './src/openMarket/index',
   ],
