+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.electron.js
@@ -11,12 +11,12 @@ import baseConfig from './webpack.config.base';
 export default validate(merge(baseConfig, {
   devtool: 'source-map',
 
-  entry: ['babel-polyfill', './app/main.development'],
+  entry: ['babel-polyfill', './src/openMarket/user_interface/main.development'],
 
   // 'main.js' in root
   output: {
     path: __dirname,
-    filename: '../../../../../app/main.js'
+    filename: '../../../user_interface/main.js'
   },
 
   plugins: [
