+++ src/openMarket/user_interface/app.global.css
@@ -1,4 +1,4 @@
-@import "~font-awesome/css/font-awesome.css";
+@import "../../../node_modules/font-awesome/css/font-awesome.css";
 
 body {
   position: relative;
