+++ src/openMarket/user_interface/components/Counter.css
