+++ src/openMarket/user_interface/components/Counter.js
