+++ src/openMarket/user_interface/reducers/counter.js
