+++ src/openMarket/user_interface/reducers/index.js
