+++ src/openMarket/user_interface/resources/icon.ico
