+++ src/openMarket/user_interface/resources/icon.png
