+++ src/openMarket/user_interface/resources/icons/1024x1024.png
