+++ src/openMarket/user_interface/resources/icons/24x24.png
