+++ src/openMarket/user_interface/resources/icons/48x48.png
