+++ src/openMarket/user_interface/resources/icons/512x512.png
