+++ src/openMarket/user_interface/resources/icons/96x96.png
