+++ src/openMarket/user_interface/store/configureStore.development.js
@@ -38,7 +38,7 @@ export default function configureStore(initialState?: counterStateType) {
 
   if (module.hot) {
     module.hot.accept('../reducers', () =>
-      store.replaceReducer(require('../reducers')) // eslint-disable-line global-require
+      store.replaceReducer(require('../reducers/index')) // eslint-disable-line global-require
     );
   }
