+++ src/test/user_interface/actions/counter.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-expressions */
 import { expect } from 'chai';
 import { spy } from 'sinon';
-import * as actions from '../../app/actions/counter';
+import * as actions from '../../../openMarket/user_interface/actions/counter';
 
 
 describe('actions', () => {
