+++ src/test/user_interface/components/Counter.spec.js
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { spy } from 'sinon';
 import React from 'react';
 import { shallow } from 'enzyme';
-import Counter from '../../app/components/Counter';
+import Counter from '../../../openMarket/user_interface/components/Counter';
 
 
 function setup() {
